Modification Added in Dashboard

+++ NainaBoutique.Models/ApplicationUser.cs
@@ -19,7 +19,7 @@ public class ApplicationUser : IdentityUser
         public int MobileNumber { get; set; }
         [NotMapped]
         public string? Role { get; set; }
-        public decimal WalletBalance { get; set; }
+        public float WalletBalance { get; set; }
 
 
     }
