+++ NainaBoutique.Models/Models/CouponModel.cs
@@ -14,13 +14,13 @@ namespace NainaBoutique.Models.Models
         public DateTime ValidTill { get; set; }
         [Required]
         [Range(1, 1000)]
-        public decimal Discount { get; set; }
+        public float Discount { get; set; }
         [Required]
         [Range(1, 1000)]
-        public decimal MinCartAmount { get; set; }
+        public float MinCartAmount { get; set; }
         [Required]
         [Range(1, 1000)]
-        public decimal MaxAmount { get; set; }
+        public float MaxAmount { get; set; }
 
 
     }
